backend/pyproject.toml
[project]
name = "liftproof-backend"
version = "0.1.0"
description = "LiftProof statistical engine - geo-testing incrementality analysis"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.109",
    "uvicorn>=0.27",
    "supabase>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    "cvxpy>=1.4",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "reportlab>=4.0",
    "sse-starlette>=1.6",
    "httpx>=0.27",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py39"
